# app/back/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.01"
  title: App
  description: API Documentation
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
#host: Taked automatically from site
basePath: /
schemes:
  # tip: remove http to make production-grade
  - http
  #- https
consumes:
  - application/json
produces:
  - application/json
paths:
  /api/private/providers:
    # binds a127 app logic to a route
    x-swagger-router-controller: provider.controller
    get:
        description: Returns the list of all Providers of user
        operationId: providers
        tags: [Providers]
        security:
          - Bearer: []
        consumes:
          - application/json
        produces:
          - application/json
        responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/providersResponse"
            # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/basicResponse"
    post: 
        description: "Acces token provider."
        operationId: addProvider
        tags: [Providers]
        security:
          - Bearer: []
        consumes:
          - application/json
        produces:
          - application/json
        parameters: 
          - name: provider
            in: body
            description: acces token provider
            required: true
            schema:
              $ref: '#/definitions/providerRequest'
        responses:
          "200":
            description: Provider token access
            schema:
             # a pointer to a definition
              $ref: "#/definitions/basicResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/basicResponse"
...
  /api/public/face/person:
    x-swagger-router-controller: face.controller
    post:
        description: Add face Recognition to Azure and Mongo
        operationId: addAzurePersonAndMongoFace
        tags: [Face, Recognition]
        consumes:
          - multipart/form-data
        produces:
          - application/json
        parameters: 
          - in: formData
            name: file
            type: file
            description: The picture to upload.
          - in: formData
            name: email
            type: string
            description: User email.
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/faceIdResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/basicResponse"
  /api/public/face/detect:
    x-swagger-router-controller: face.controller
    post:
        description: Get Azure faceId
        operationId: getAzureFaceId
        tags: [Face, Recognition]
        consumes:
          - multipart/form-data
        produces:
          - application/json
        parameters: 
          - in: formData
            name: file
            type: file
            description: The picture to upload.
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/faceIdResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/basicResponse"
...
definitions:
  providerRequest:
    type: object
    properties:
      code:
        type: string
        example: code client
      provider:
        type: string
        example: provider type
  providersResponse:
    required:
      - list
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/providerResponse'
  providerResponse:
    type: object
    properties:
      _id:
        type: string
        example: Todo Id
      title:
        type: string
        example: Title of To Do
      description:
        type: string
        example: Desciption of To Do
      user: 
        type: string
        example: User Id
      created:
        type: string
        example: Date of To Do Created
...
  basicResponse:
    type: object
    properties:
      message:
        type: string
  faceIdResponse:
    type: object
    properties:
      faceId:
        type: string
...